backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.7
email-validator>=2.1.0
motor>=3.3.0
dnspython>=2.6.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
cloudinary>=1.39.0
requests>=2.31.0  
pydantic-settings>=2.2.0
opencv-contrib-python-headless>=4.9.0.80
pillow>=10.2.0
pillow-heif>=0.15.0
torch>=2.2.0
torchvision>=0.17.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.12.0
matplotlib>=3.8.0
bcrypt==4.0.1